style(campaign): fix clippy lints in stage3 interactive harness
Controller-applied exception (worker run timed out twice): 3 mechanical
clippy fixes in test-only files — never_loop (select arms both exit),
too_many_arguments allow on audit helper, unnecessary usize cast.
Behavior unchanged; campaign_stage3_interactive 6/6 green.

+++ src/web/workspace_ws_handler/tests/campaign_stage3_interactive/cases.rs
@@ -249,7 +249,7 @@ async fn campaign_stage3_oversized_feedback_rejects_before_turn_reservation() {
 
     // 案 1：反馈文本超长。
     let oversized_feedback = "x".repeat(
-        crate::product::workspace_engine::SC_MANUAL_REVISION_FEEDBACK_MAX_BYTES as usize + 1,
+        crate::product::workspace_engine::SC_MANUAL_REVISION_FEEDBACK_MAX_BYTES + 1,
     );
     let before = harness.session_bytes();
     harness
